+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.0.0
   - Breaking: restrict unintended user constant usage [#56](https://github.com/logstash-plugins/logstash-filter-ruby/pull/56)
+  - Fix: do not handle potential fatal (Java) errors - let them propagate out
 
 ## 3.1.7
   - [DOC] Added docs to help people avoid concurrency issues (often caused by accidentally relying on shared state with global variables, constants, or unguarded overwriting of instance variables) [#58](https://github.com/logstash-plugins/logstash-filter-ruby/issues/58)